dt는 float이다. 거기에 정밀하게 속도를 곱해주는데 _gone은 int 형임. 더구나 dt는 아주 작은 값임. 이로 인하여 제대로 누적이 되지 않음 화면 한페이지당 6픽셀 이상 밀림.
발견하기 어려웠다.

+++ Flappy Bird/GameLayer.h
@@ -25,13 +25,13 @@
 @property (nonatomic) BOOL gameOver;
 @property (nonatomic) int screenSpeed;
 @property (nonatomic) int delayPipeStart;
-@property (nonatomic) int pipeGap;
+@property (nonatomic) float pipeGap;
 @property (nonatomic) int score;
-@property (nonatomic) int gone;
+@property (nonatomic) float gone;
 @property (nonatomic) CGSize winSize;
 @property (nonatomic) int  pipeUpDownGap;
 
-@property (nonatomic) int birdHeight;
+@property (nonatomic) float birdHeight;
 @property (nonatomic) float velocity;
 @property (nonatomic) float impactTime;
